Incorporate code generators to websub compiler plugin
